Add files via upload
This semester, I started learning Python for school. Writing code that actually works is one thing but improving it is real hard! My task was to optimize a program that chat gpt wrote. 

This idea comes from Lagrange’s Four-Square Theorem, which says every whole number can be written as the sum of four squares.
n = a^2 + b^2 + c^2 + d^2

Example:
9 = 3² + 0² + 0² + 0²
30 = 5² + 2² + 1² + 0²


How it works!:

If the number n happens to be a perfect square, the solution is immediate. Example: n = 9 is [3² + 0² + 0² + 0²]

But...
Otherwise, it searches smartly. While reading about optimization, I discovered Legendre’s Three-Square Theorem, which gives an incredibly useful shortcut: 

A positive integer can be written as the sum of three squares if and only if it’s not of the form: 

n = 4^k(8x+7)
translated: If the number is divisible by 4, then divide it until it is no longer divisible by 4. And if that number gives a remainder of 7 when divided by 8, then it’s impossible to write your original n as the sum of three squares.

In this case, I will make this n = r, where r = n - a². I then made a loop to test values of a that satisfy r where a cannot be greater than the square root of n.

After applying Legendre’s shortcut, the program looks at what’s left. It then tests different values for b^2 and subtracts b^2 from r to get a smaller number. Finally, it checks if that leftover can be written as the sum of two more squares, c^2+d^2


This project taught me so much and I'm excited to keep improving!